% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fmdu.R
\name{plot.fmdu}
\alias{plot.fmdu}
\title{Visualisation of a fmdu objects}
\usage{
\method{plot}{fmdu}(x, ...)
}
\arguments{
\item{x}{An fmdu object .}

\item{...}{additional arguments to pass}
}
\value{
No return value, called for side effects (plot)
}
\description{
Plot method for a \code{fmdu} object. The plot shows the result of \code{fmdu}.
}
