% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fmdu.R
\name{print.fmdu}
\alias{print.fmdu}
\title{Print method for all fmdu objects}
\usage{
\method{print}{fmdu}(x, ...)
}
\arguments{
\item{x}{object of class \code{fmdu}.}

\item{\dots}{additional arguments to be passed.}
}
\value{
No return value, called for side effects (print)
}
\description{
Print the results of a fmdu object
}
