% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fmdu.R
\name{summary.fmdu}
\alias{summary.fmdu}
\title{Summary method for all fmdu objects}
\usage{
\method{summary}{fmdu}(object, ...)
}
\arguments{
\item{object}{object of class \code{fmdu}.}

\item{\dots}{additional arguments to be passed.}
}
\value{
No return value, called for side effects (summary)
}
\description{
Summary method for all fmdu objects
}
